Fixed save method in the account classes, to now have one method for saving with encoding, and one without.(Problem where user could not log in because password ggot encoded after adding favvourites)

+++ src/main/java/no/ntnu/entity/models/Cars.java
@@ -195,7 +195,7 @@ public class Cars {
     return passengers;
   }
 
-  public void setPassengers(byte passengers) {
+  public void setPassengers(int passengers) {
     this.passengers = passengers;
   }
